+++ test/openBCIUtilities-test.js
@@ -917,7 +917,6 @@ describe('openBCIUtilities', function () {
   });
   describe('#makeDaisySampleObject', function () {
     let lowerSampleObject, upperSampleObject, daisySampleObject;
-    let lowerSampleObjectNoScale, upperSampleObjectNoScale, daisySampleObjectNoScale;
     before(() => {
       // Make the lower sample (channels 1-8)
       lowerSampleObject = openBCIUtilities.newSample(1);
@@ -932,59 +931,35 @@ describe('openBCIUtilities', function () {
       upperSampleObject.timestamp = 8;
 
       daisySampleObject = openBCIUtilities.makeDaisySampleObject(lowerSampleObject, upperSampleObject);
-
-      // lowerSampleObjectNoScale = openBCIUtilities.newSample(1);
-      // lowerSampleObjectNoScale.channelDataCounts = [1, 2, 3, 4, 5, 6, 7, 8];
-      // lowerSampleObjectNoScale.auxData = [0, 1, 2];
-      // lowerSampleObjectNoScale.timestamp = 4;
-      // lowerSampleObjectNoScale.accelData = [0.5, -0.5, 1];
-      // // Make the upper sample (channels 9-16)
-      // upperSampleObjectNoScale = openBCIUtilities.newSample(2);
-      // upperSampleObjectNoScale.channelDataCounts = [9, 10, 11, 12, 13, 14, 15, 16];
-      // upperSampleObjectNoScale.auxData = [3, 4, 5];
-      // upperSampleObjectNoScale.timestamp = 8;
-      //
-      // // Call the function under test
-      // daisySampleObjectNoScale = openBCIUtilities.makeDaisySampleObject(lowerSampleObjectNoScale, upperSampleObjectNoScale);
     });
     it('should make a channelData array 16 elements long', function () {
       expect(daisySampleObject.channelData).to.have.length(k.OBCINumberOfChannelsDaisy);
-      // expect(daisySampleObjectNoScale.channelDataCounts).to.have.length(k.OBCINumberOfChannelsDaisy);
     });
     it('should make a channelData array with lower array in front of upper array', function () {
       for (let i = 0; i < 16; i++) {
         expect(daisySampleObject.channelData[i]).to.equal(i + 1);
-        // expect(daisySampleObjectNoScale.channelDataCounts[i]).to.equal(i + 1);
       }
     });
     it('should make the sample number equal to the second packet divided by two', function () {
       expect(daisySampleObject.sampleNumber).to.equal(upperSampleObject.sampleNumber / 2);
-      // expect(daisySampleObjectNoScale.sampleNumber).to.equal(upperSampleObject.sampleNumber / 2);
     });
     it('should put the aux packets in an object', function () {
       expect(daisySampleObject.auxData).to.have.property('lower');
       expect(daisySampleObject.auxData).to.have.property('upper');
-      // expect(daisySampleObjectNoScale.auxData).to.have.property('lower');
-      // expect(daisySampleObjectNoScale.auxData).to.have.property('upper');
     });
     it('should put the aux packets in an object in the right order', function () {
       for (let i = 0; i < 3; i++) {
         expect(daisySampleObject.auxData['lower'][i]).to.equal(i);
         expect(daisySampleObject.auxData['upper'][i]).to.equal(i + 3);
-        // expect(daisySampleObjectNoScale.auxData['lower'][i]).to.equal(i);
-        // expect(daisySampleObjectNoScale.auxData['upper'][i]).to.equal(i + 3);
       }
     });
     it('should average the two timestamps together', function () {
       let expectedAverage = (upperSampleObject.timestamp + lowerSampleObject.timestamp) / 2;
       expect(daisySampleObject.timestamp).to.equal(expectedAverage);
-      // expect(daisySampleObjectNoScale.timestamp).to.equal(expectedAverage);
     });
     it('should place the old timestamps in an object', function () {
       expect(daisySampleObject._timestamps.lower).to.equal(lowerSampleObject.timestamp);
       expect(daisySampleObject._timestamps.upper).to.equal(upperSampleObject.timestamp);
-      // expect(upperSampleObjectNoScale._timestamps.lower).to.equal(upperSampleObjectNoScale.timestamp);
-      // expect(upperSampleObjectNoScale._timestamps.upper).to.equal(upperSampleObjectNoScale.timestamp);
     });
     it('should store an accelerometer value if present', function () {
       expect(daisySampleObject).to.have.property('accelData');
@@ -2009,6 +1984,105 @@ describe('#extractRawDataPackets', function () {
  * Test the function that routes raw packets for processing
  */
 describe('#transformRawDataPacketsToSamples', function () {
+  it('should process three ganglion packets packet', function () {
+    let rawDataToSample = k.rawDataToSampleObjectDefault(k.OBCINumberOfChannelsGanglion);
+    rawDataToSample.protocol = k.OBCIProtocolWifi;
+    rawDataToSample.lastSampleNumber = 0;
+
+    rawDataToSample.rawDataPackets = [
+      openBCIUtilities.samplePacket(0),
+      openBCIUtilities.samplePacket(1),
+      openBCIUtilities.samplePacket(2)
+    ];
+
+    // gain here is 51, the same as in the channel settings array
+    let scaleFactor = 1.2 / 51 / (Math.pow(2, 23) - 1);
+
+    // Call the function under test
+    let samples = openBCIUtilities.transformRawDataPacketsToSample(rawDataToSample);
+
+    for (let j = 0; j < k.OBCINumberOfChannelsGanglion; j++) {
+      // console.log(`channel data ${j + 1}: ${valueArray[j]} : actual ${scaleFactor * (j + 1)}`)
+      expect(samples[0].channelData[j]).to.be.closeTo(scaleFactor * (j + 1), 0.00000001);
+      expect(samples[1].channelData[j]).to.be.closeTo(scaleFactor * (j + 1), 0.00000001);
+      expect(samples[2].channelData[j]).to.be.closeTo(scaleFactor * (j + 1), 0.00000001);
+    }
+
+    // Ensure that we extracted only one buffer
+    expect(samples).to.have.length(3);
+  });
+  it('should process three cyton packets packet', function () {
+    let rawDataToSample = k.rawDataToSampleObjectDefault(k.OBCINumberOfChannelsCyton);
+    rawDataToSample.protocol = k.OBCIProtocolWifi;
+    rawDataToSample.lastSampleNumber = 0;
+
+    rawDataToSample.rawDataPackets = [
+      openBCIUtilities.samplePacket(0),
+      openBCIUtilities.samplePacket(1),
+      openBCIUtilities.samplePacket(2)
+    ];
+
+    // gain here is 24, the same as in the channel settings array
+    let scaleFactor = 4.5 / 24 / (Math.pow(2, 23) - 1);
+
+    // Call the function under test
+    let samples = openBCIUtilities.transformRawDataPacketsToSample(rawDataToSample);
+
+    for (let j = 0; j < k.OBCINumberOfChannelsDefault; j++) {
+      // console.log(`channel data ${j + 1}: ${valueArray[j]} : actual ${scaleFactor * (j + 1)}`)
+      expect(samples[0].channelData[j]).to.be.closeTo(scaleFactor * (j + 1), 0.00000001);
+      expect(samples[1].channelData[j]).to.be.closeTo(scaleFactor * (j + 1), 0.00000001);
+      expect(samples[2].channelData[j]).to.be.closeTo(scaleFactor * (j + 1), 0.00000001);
+    }
+
+    // Ensure that we extracted only one buffer
+    expect(samples).to.have.length(3);
+  });
+  it('should process three daisy packets packet', function () {
+    let rawDataToSample = k.rawDataToSampleObjectDefault(k.OBCINumberOfChannelsDaisy);
+    rawDataToSample.protocol = k.OBCIProtocolWifi;
+    rawDataToSample.lastSampleNumber = 0;
+
+    rawDataToSample.rawDataPackets = [
+      openBCIUtilities.samplePacket(0),
+      openBCIUtilities.samplePacket(1),
+      openBCIUtilities.samplePacket(1)
+    ];
+
+    for (let i = 0; i < k.OBCINumberOfChannelsDefault; i++) {
+      rawDataToSample.channelSettings[i].gain = 1;
+    }
+    // gain here is 24, the same as in the channel settings array
+    let scaleFactor = 4.5 / 24 / (Math.pow(2, 23) - 1);
+    let scaleFactor1 = 4.5 / (Math.pow(2, 23) - 1);
+
+    // Call the function under test
+    let samples = openBCIUtilities.transformRawDataPacketsToSample(rawDataToSample);
+
+    for (let j = 0; j < k.OBCINumberOfChannelsDefault; j++) {
+      // console.log(`channel data ${j + 1}: ${valueArray[j]} : actual ${scaleFactor * (j + 1)}`)
+      expect(samples[0].channelData[j]).to.be.closeTo(scaleFactor * (j + 1), 0.00000001);
+    }
+
+    for (let j = 0; j < k.OBCINumberOfChannelsDefault; j++) {
+      // console.log(`channel data ${j + 1}: ${valueArray[j]} : actual ${scaleFactor * (j + 1)}`)
+      expect(samples[1].channelData[j]).to.be.closeTo(scaleFactor1 * (j + 1), 0.00000001);
+    }
+
+    for (let j = 0; j < k.OBCINumberOfChannelsDefault; j++) {
+      // console.log(`channel data ${j + 1}: ${valueArray[j]} : actual ${scaleFactor * (j + 1)}`)
+      expect(samples[2].channelData[j]).to.be.closeTo(scaleFactor * (j + 1), 0.00000001);
+    }
+
+    // Ensure that we extracted only one buffer
+    expect(samples).to.have.length(3);
+  });
+});
+
+/**
+ * Test the function that routes raw packets for processing
+ */
+describe('#transformRawDataPacketToSample', function () {
   var funcSpyTimeSyncedAccel, funcSpyTimeSyncedRawAux, funcSpyStandardRawAux, funcSpyStandardAccel;
 
   before(function () {
@@ -2033,36 +2107,20 @@ describe('#transformRawDataPacketsToSamples', function () {
     var buffer = openBCIUtilities.samplePacket(0);
 
     // Call the function under test
-    const samples = openBCIUtilities.transformRawDataPacketsToSample({
-      rawDataPackets: [buffer],
+    openBCIUtilities.transformRawDataPacketToSample({
+      rawDataPacket: buffer,
       channelSettings: defaultChannelSettingsArray
     });
 
     // Ensure that we extracted only one buffer
     expect(funcSpyStandardAccel).to.have.been.calledOnce();
-    expect(samples.length).to.equal(1);
-  });
-
-  it('should process a standard packet', function () {
-    // Call the function under test
-    openBCIUtilities.transformRawDataPacketsToSample({
-      rawDataPackets: [
-        openBCIUtilities.samplePacket(0),
-        openBCIUtilities.samplePacket(1),
-        openBCIUtilities.samplePacket(2)
-      ],
-      channelSettings: defaultChannelSettingsArray
-    });
-
-    // Ensure that we extracted only one buffer
-    expect(funcSpyStandardAccel).to.have.been.calledThrice();
   });
   it('should process a standard packet with raw aux', function () {
     var buffer = openBCIUtilities.samplePacketStandardRawAux(0);
 
     // Call the function under test
-    openBCIUtilities.transformRawDataPacketsToSample({
-      rawDataPackets: [buffer]
+    openBCIUtilities.transformRawDataPacketToSample({
+      rawDataPacket: buffer
     });
 
     // Ensure that we extracted only one buffer
@@ -2072,8 +2130,8 @@ describe('#transformRawDataPacketsToSamples', function () {
     var buffer = openBCIUtilities.samplePacketUserDefined();
 
     // Call the function under test
-    openBCIUtilities.transformRawDataPacketsToSample({
-      rawDataPackets: [buffer]
+    openBCIUtilities.transformRawDataPacketToSample({
+      rawDataPacket: buffer
     });
 
     // Nothing should be called
@@ -2086,8 +2144,8 @@ describe('#transformRawDataPacketsToSamples', function () {
     var buffer = openBCIUtilities.samplePacketAccelTimeSyncSet();
 
     // Call the function under test
-    openBCIUtilities.transformRawDataPacketsToSample({
-      rawDataPackets: [buffer]
+    openBCIUtilities.transformRawDataPacketToSample({
+      rawDataPacket: buffer
     });
 
     // we should call to get a packet
@@ -2097,8 +2155,8 @@ describe('#transformRawDataPacketsToSamples', function () {
     var buffer = openBCIUtilities.samplePacketAccelTimeSynced(0);
 
     // Call the function under test
-    openBCIUtilities.transformRawDataPacketsToSample({
-      rawDataPackets: [buffer]
+    openBCIUtilities.transformRawDataPacketToSample({
+      rawDataPacket: buffer
     });
 
     // Ensure that we extracted only one buffer
@@ -2108,8 +2166,8 @@ describe('#transformRawDataPacketsToSamples', function () {
     var buffer = openBCIUtilities.samplePacketRawAuxTimeSyncSet(0);
 
     // Call the function under test
-    openBCIUtilities.transformRawDataPacketsToSample({
-      rawDataPackets: [buffer]
+    openBCIUtilities.transformRawDataPacketToSample({
+      rawDataPacket: buffer
     });
 
     expect(funcSpyTimeSyncedRawAux).to.have.been.calledOnce();
@@ -2118,8 +2176,8 @@ describe('#transformRawDataPacketsToSamples', function () {
     var buffer = openBCIUtilities.samplePacketRawAuxTimeSynced(0);
 
     // Call the function under test
-    openBCIUtilities.transformRawDataPacketsToSample({
-      rawDataPackets: [buffer]
+    openBCIUtilities.transformRawDataPacketToSample({
+      rawDataPacket: buffer
     });
 
     // Ensure that we extracted only one buffer
@@ -2132,8 +2190,8 @@ describe('#transformRawDataPacketsToSamples', function () {
     buffer[k.OBCIPacketPositionStopByte] = 0xCF;
 
     // Call the function under test
-    openBCIUtilities.transformRawDataPacketsToSample({
-      rawDataPackets: [buffer]
+    openBCIUtilities.transformRawDataPacketToSample({
+      rawDataPacket: buffer
     });
 
     // Nothing should be called